refactor: 중복된 역할과 책임을 가지는 클래스 제거
