+++ src/main/java/chess/domain/state/Running.java
@@ -2,11 +2,6 @@ package chess.domain.state;
 
 import chess.domain.board.Board;
 import chess.domain.board.Position;
-import chess.domain.board.Score;
-import chess.domain.board.ScoreResult;
-import chess.domain.piece.Color;
-import java.util.HashMap;
-import java.util.Map;
 
 abstract class Running extends Ready {
